Remove unnecessary public wrapper types
We don't need to expose Lexer and Parser, as they are effectively
implementation details. What's interesting is the data they provide:
Tokens, and the AST, both of which are already exposed.

Move to using a free-standing Lex() and Parse() function which returns
the info necessary instead, and hide the types.