Separate keywords, identifiers
We add a new token (TokenKeyword) to represent all reserved identifiers,
and use it to find things in the right place. We use the newly added
'sequence' keyword, and also add a
TokenNamespace, TokenOpenSquareBracket, TokenCloseSquareBracket,
TokenGreaterThan, TokenLessThan --
and use all of these to implement a proper 'parseType' that somewhat
approximates reasonable behaviour.

This enables us to introduce better handling of types rather than just
doing dumb things like reading everything up to a comma (which is wrong,
in the case of sequence):

    sequence<long, 10>
    sequence<long,10>

... should be treated identically, but we were previously failing on the
first case.

Additionally, introduce a parseValue(), to keep:

    const Foo Bar = 0x0001 << 0;

... working, despite the new LT and GT tokens.

Type handling is still not good enough:

    string
    string<10>
    string<Foo>
    string<Foo::Bar>

... all need to work. But this is a lot closer. Perhaps the sequence
keyword is a mistake...